requirements.txt
streamlit
gTTS
SpeechRecognition
requests
schedule
googlesearch-python
plyer